Licensed under Creative Commons BY-NC-SA
~~ Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International Public License ~~
Simple explanation of the license terms: https://creativecommons.org/licenses/by-nc-sa/4.0/

I chose this license to permit anyone to be able to read this work, or to submit corrections, modifications, or improvements. If anyone creates a work based on this one, since this is intended to be built upon by and for the community, I require that derivative work to also have the same open-source licence. Similarly, since I do not intend to gain any revenue from publishing this work, I am enforcing a non-commercial clause. For any use, the main authors and contributors must be properly attributed for their efforts.